# frontend/openapi.json
{"openapi":"3.0.0","paths":{"/auth/google/url":{"get":{"operationId":"AuthController_getGoogleAuthUrl","summary":"Get Google OAuth authorization URL","parameters":[],"responses":{"200":{"description":"Returns the OAuth URL"}},"tags":["Authentication"]}},"/auth/google/callback":{"get":{"operationId":"AuthController_handleGoogleCallback","summary":"Handle Google OAuth callback","parameters":[],"responses":{"302":{"description":"Redirects to frontend"}},"tags":["Authentication"]}},"/auth/logout":{"post":{"operationId":"AuthController_logout","summary":"Logout current user","parameters":[],"responses":{"200":{"description":"Session revoked"}},"tags":["Authentication"]}},"/auth/me":{"get":{"operationId":"AuthController_getCurrentUser","summary":"Get current authenticated user","parameters":[],"responses":{"200":{"description":"Returns the current user"},"401":{"description":"Unauthorized"}},"tags":["Authentication"]}},"/health":{"get":{"operationId":"HealthController_check","summary":"Health check","parameters":[],"responses":{"200":{"description":"Service is healthy"}},"tags":["Health"]}},"/applications":{"post":{"operationId":"ApplicationController_create","summary":"Create a new application","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateApplicationDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationResponseDto"}}}}},"tags":["Applications"],"security":[{"bearer":[]}]},"get":{"operationId":"ApplicationController_findAll","summary":"List all applications","parameters":[{"name":"status","required":true,"in":"query","schema":{"type":"string"}},{"name":"source","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationListResponseDto"}}}}},"tags":["Applications"],"security":[{"bearer":[]}]}},"/applications/timeline":{"get":{"operationId":"ApplicationController_getTimeline","summary":"Get application timeline grouped by date","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TimelineResponseDto"}}}}},"tags":["Applications"],"security":[{"bearer":[]}]}},"/applications/{id}":{"get":{"operationId":"ApplicationController_findById","summary":"Get application by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationResponseDto"}}}},"404":{"description":"Not found"}},"tags":["Applications"],"security":[{"bearer":[]}]}},"/applications/{id}/status":{"patch":{"operationId":"ApplicationController_updateStatus","summary":"Update application status","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateApplicationStatusDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationResponseDto"}}}}},"tags":["Applications"],"security":[{"bearer":[]}]}},"/email-sync/sync":{"post":{"operationId":"EmailSyncController_startSync","summary":"Start initial email sync from Gmail","parameters":[],"responses":{"201":{"description":"Sync completed"}},"tags":["Email Sync"],"security":[{"bearer":[]}]}}},"info":{"title":"CareerOS API","description":"CareerOS backend API","version":"0.1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"cookie":{"type":"apiKey","in":"cookie","name":"session"}},"schemas":{"CreateApplicationDto":{"type":"object","properties":{"jobId":{"type":"string"},"companyName":{"type":"string"},"jobTitle":{"type":"string"},"notes":{"type":"string"},"appliedAt":{"type":"string"},"source":{"type":"string","enum":["manual","email"]}}},"ApplicationResponseDto":{"type":"object","properties":{"id":{"type":"string"},"status":{"type":"string"},"source":{"type":"string"},"jobId":{"type":"string"},"jobTitle":{"type":"string"},"companyName":{"type":"string"},"notes":{"type":"string"},"appliedAt":{"type":"string"},"createdAt":{"type":"string"},"updatedAt":{"type":"string"}},"required":["id","status","source","createdAt","updatedAt"]},"ApplicationListResponseDto":{"type":"object","properties":{"applications":{"type":"array","items":{"$ref":"#/components/schemas/ApplicationResponseDto"}},"total":{"type":"number"}},"required":["applications","total"]},"TimelineEntryDto":{"type":"object","properties":{"date":{"type":"string"},"applications":{"type":"array","items":{"$ref":"#/components/schemas/ApplicationResponseDto"}}},"required":["date","applications"]},"TimelineResponseDto":{"type":"object","properties":{"timeline":{"type":"array","items":{"$ref":"#/components/schemas/TimelineEntryDto"}}},"required":["timeline"]},"UpdateApplicationStatusDto":{"type":"object","properties":{"status":{"type":"string","enum":["Draft","Applied","Screening","Interviewing","Offer","Accepted","Rejected","Declined","Closed"]}},"required":["status"]}}}}